% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Implementation of the brute_force_knapsack function.}
\arguments{
\item{x}{A data frame that contains two positive vectors for weights (w) and values (v).}

\item{W}{The maximum knapsack capacity.}

\item{parallel}{set to FALSE by default.}
}
\value{
A list containing the values and elements used to fill the knapsack.
}
\description{
The following code implements the knapsack Brute Force Search algorithm.
}
\examples{
set.seed(42)
n <- 2000
w = sample(1:4000, size = n, replace = TRUE)
v = runif(n = n, 0, 10000)
brute_force_knapsack(x = knapsack_objects[1:8,], W = 3500)
brute_force_knapsack(x = knapsack_objects[1:12,], W = 3500)
brute_force_knapsack(x = knapsack_objects[1:8,], W = 2000)
brute_force_knapsack(x = knapsack_objects[1:12,], W = 2000)
}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack_problem}
}
