% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.r
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Implementation of the "0-1" knapsack_dynamic function.}
\arguments{
\item{x}{A data frame that contains two positive vectors for weights (w) and values (v).}

\item{W}{The maximum knapsack capacity.}
}
\value{
A list containing the values and elements used to fill the knapsack.
}
\description{
The following algorithm calculates m(n,W). A more detailed explanation of the pseudo-algorithm followed can be found on Wikipedia.
}
\examples{
set.seed(42)
n <- 2000
w = sample(1:4000, size = n, replace = TRUE)
v = runif(n = n, 0, 10000)
knapsack_objects <- data.frame(w,v)
knapsack_dynamic(x = knapsack_objects[1:8,], W = 3500)
knapsack_dynamic(x = knapsack_objects[1:12,], W = 3500)
knapsack_dynamic(x = knapsack_objects[1:8,], W = 2000)
knapsack_dynamic(x = knapsack_objects[1:12,], W = 2000)
}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack_problem}
}
